refactor: Move `ExecutableError` to handler to avoid circular dep
